requirements.txt
pandas
numpy
scikit-learn
tensorflow
pyyaml
matplotlib
seaborn